rearranged extension classes into the Extensions namespace and moved enum-based extensions to the EnumExtensions class. closes #4
